requirements.txt
sounddevice
deepgram-sdk 
websockets==12.0
elevenlabs==0.2.26
openai 
python-dotenv 
pydub

grpcio==1.51.3
numpy==1.22.4
soundfile==0.12.1
protobuf==3.20.0
mediapipe==0.10.11
flask==2.3.3
requests==2.31.0
pyaudio